Made the default URL be the `only-qr-codes` path created in the last commit.
...I think. It looks like that's what I did, but I haven't tested it!
Since the trackers aren't currently working as deployed, this won't be a
regression in practice, but rest assured that I will try to make a mode
that gives the best of both worlds, both trackers and QR codes. But
right now I just want to see the fruits of my labor :).