amend! Working QR Codes (#19)
Working QR Codes (#19)

* Caused deployed copy to update from this branch in my fork...

...only when deployed from this repo.

* Updated password, added my ssh key.

Yay! Now I can log in!

* Made users.mutableUsers = true explicit.

It wasn't before, which caused me a headache.

* Installed ffmpeg.

This enables my hilarious workflow of developing remotely using the VPN, taking a video of the screen with the webcam using
ffmpeg, and looking at that to see what I've done. This idea was all Chase's, and it is glorious.

* It works!

It was caching on the browser. "Ctrl+F5" fixed it... :(

* Update qrCodeUrls.ts

Updated LUG's Discord QR Code

* amend! It works!

It works!

It was caching on the browser. "Ctrl+F5" fixed it... :(

amendment: This implements QR codes to the Discords and the Github, and
makes them the page shown to the user by default.

* Actually clearing browser cache on deploy.

The browser cache was not being cleared on deploy, despite our use of
`--disable-http-cache` when invoking chrome. I have removed that flag
because I believe it does nothing (rationale below), and instead I am
causing chrome's cache directory to be removed on every deploy with `rm
-r ~/.cache/google-chrome`. This appears new technique with `rm` appears
to avoid exactly the kinds of stale data problems that bugged me so - I
tested this somewhat exhaustively, and `rm -rf ~/.cache/google-chrome`'s
presence or absence was perfectly predictive of whether a change to a qr
code's url would be reflected or not, respectively, in the next deploy.
(I did not test further deploys).

I believe `--disable-http-cache` not to be a real flag because:

1. It is not listed on
https://peter.sh/experiments/chromium-command-line-switches/, the best
list of Google Chrome CLI flags that I can find. All the other flags we
use are listed here.

2. Chrome appears happy to launch with any invalid flag. I've tried this
by modifying our `nix/system.nix` to have the flags
`--disable-http-cach` (intentional misspelling) and another hypothetical
flag that was inappropriate, and deployment went off without a hitch.
This is shocking and disappointing. I haven't bothered to find if this
is otherwise documented, and I haven't looked in any logs for
complaints, mostly because I wouldn't know exactly where to look.

3. It doesn't actually fix our stale-data problem, which `Ctrl+F5` on an
attached keyboard (triggering a hard refresh) *does* fix.

* Updating from the `acm-uic` repository again.

* No longer using `display: inline` for qr-codes.

As per @Ajknight121's recommendation to reserve `display: inline` for
text elements.

amendment 2: I have confirmed that Google Chrome does not care what
invalid flags you give it by running `Version 140.0.7339.186 (Official
Build) (arm64)` on macOS with a very profane flag.
---------

Co-authored-by: Michael Khan <michaelk30200389@gmail.com>