Trivial Promises -> async/await.
It's rare that something just works to make code shorter like this. It's a good feeling.

I don't think we should factor out the common elements of all this code. Either we should compress the "get" functions
into the controllers, or we should put all the controllers (and get functions?) together, or both. All of those are not
very enterprisey but hey, I think they'd be better.